safewatch_ai/requirements.txt
ultralytics>=8.0.0,<9.0.0
opencv-python>=4.8.1
numpy>=1.24.0,<2.0
streamlit>=1.28.0
pandas>=2.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
email-validator>=2.0.0
